trial with test_cli_runner
