+++ test/tools/test_utils.h
@@ -31,19 +31,24 @@ std::string read_file(FILE* file);
 // Compares the contents of two FILE* streams.
 bool compare_files(FILE* expected, FILE* actual);
 
+// Prints the line-by-line diff between expected and actual output.
+void print_diff(const std::string& expected, const std::string& actual);
+
+// Replaces placeholder paths like $EXEEXT and $DATA_DIR in command strings.
+std::string adjust_tool_path(const std::string& raw_command);
+
 // Runs a single test case and populates the result object.
 // Returns 1 if the test fails, 0 otherwise.
-int run_test(const std::string& id,
-             const std::string& cmd,
+int run_test(const std::string& cmd,
              FILE* expected_stdout,
              int expected_exit,
-             TestResult& result,
-             std::string tmpdir);
+             TestResult& result);
 
 // Prints a summary of all test results to stdout.
 void print_summary(const std::vector<TestResult>& results);
 
-// Loads and runs all tests from the cli_tests.txt file.
+// Loads and runs all tests from cli_tests.txt.
 // Returns 0 on success (all tests passed), 1 otherwise.
-int run_all_tests(std::string tmpdir);
+int run_all_tests();
+
 #endif // TSK_CLI_RUNNER_H
